testutils: Consolidate DummySelectorCacheUser
No functional changes in this commit, just moving code.
