+++ pkg/policy/rule.go
@@ -17,6 +17,7 @@ import (
 	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/option"
 	"github.com/cilium/cilium/pkg/policy/api"
+	policytypes "github.com/cilium/cilium/pkg/policy/types"
 )
 
 // ruleKey is a synthetic unique identifier for a Rule
@@ -39,7 +40,7 @@ type rule struct {
 // IdentitySelectionUpdated is called by the SelectorCache when a new identity is added;
 // We can ignore it because the endpoint will be regenerated by the nature of
 // identities being updated.
-func (r *rule) IdentitySelectionUpdated(_ CachedSelector, _, _ []identity.NumericIdentity) {
+func (r *rule) IdentitySelectionUpdated(_ policytypes.CachedSelector, _, _ []identity.NumericIdentity) {
 }
 
 func (d *rule) IdentitySelectionCommit(*versioned.Tx) {
